Create Select Screen
Putting all these in the /src folder for a React app should
render a page where you can use buttons to navigate from
page to page. This is a little unintuitive because the back
arrows won't change your page, as there's technically only one
page. That could be a performance issue as well, as it seemed
like the page could take a little while to compile at times.
Feel free to change the implementation, but I wanted something
quick to more easily divide labor while sharing certain utilities,
such as the turtle class.