Functions to manipulate turtle collection
You can now use getTurtles() and unlockTurtle() to get a user's
array of turtles and set one of the elements in the array to true,
respectively. In my mind, the collection of turtles should be
stored as an array of booleans, with each index corresponding to
a given turtle being unlocked/purchased for the user.