Sign-Out Button
You can put a Sign-Out-Button anywhere you want now. I'm a bit
concerned about the ability to navigate back to a previous page
though, so we'll need to make sure a user can't access the proper
features if not logged in.
A small note: if you're logging in for the first time, doing so
through the debugging window opened by VSCode can raise some
eyebrows with Google. Just copy the url into chrome elsewhere, and
signing in should be fine. After that, I think using your debug
window works well.