Realtime database proof of concept
Firestore realtime database can be used to render a page which
will re-render every time there's a change in particular variables
in the database. Using this, it seems feasible that an opponent and
a player could view screens with the same information, but perhaps
mirrored. The previous work on the game logic is preserved in
game-cycle-old.jsx.