Game that can be played in two tabs
We're getting somewhat close now. I put in two links, one will
take you to the red "team" and the other will take you to the
"blue" team. The pages have different views but are perfectly in
sync. The next step is to create the mechanism by which only two
users will be able to be in a room at a given time.