Exclude users who are not signed in
I'm actually very excited about this change. There's now a clear
way to ensure that only signed-in users can view the pages we
want, and it's very clear how to get the users email and uid in
your component: pass it via props in index.js. On a side note,
I've noticed a bug in the attack buttons where every so often one
of them won't trigger the onClick properly. I'm not sure what
that's about, but hopefully it becomes more clear.