Users get a profile in the database!
Very excited for this as well :). Upon logging in, if user's have
a profile (associated with their UID), then their login date will
be stored in the database. If they don't have a profile, a brand
new one will be created. Go ahead and try it! Currently default
is you have the first turtle unlocked, none of them are named,
and you have 100 currency in your wallet.