Login bonuses :)
If you log in on the date after the date that you last played,
you'll have 100 coins added to your wallet. To demo this, simply
manually change your login date to the previous date, and then
sign in. NOTE: the date is given in day/month/year format, for
better or for worse.