Search for a user and display their profile
My new working plan is to have the challenger search for the
username of the player they would like to challenge. The
challenger will get a nice layout of the player's page, and
can then choose to send them a challenge. Ideally, this
challenge will appear at the bottom of the home screen of the
player being challenged.