Working online??
Due in large to the efforts of Lucas, we finally have a system
that should be able to connect to users via firebase. It's not
without bugs though: I still need to fix the ability to play
the same move as your opponent, and the win messages can be a
little funky. It also seems like the room doesn't get deleted
afterwards, like one would expect.